Fix several issues reported
CBL-2088 (getDocIDs HTTP 500), CBL-2087 (replicator_config API), password not being honored, build script not failing on failed build

+++ CBLClient/Apps/CBLTestServer-C/src/ReplicatorConfigurationMethods.cpp
@@ -226,7 +226,7 @@ namespace replicator_configuration_methods {
             }
 
             if(body.contains("heartbeat")) {
-                config->heartbeat = stoul(body["heartbeat"].get<string>(), nullptr, 10);
+                config->heartbeat = (unsigned int)stoul(body["heartbeat"].get<string>(), nullptr, 10);
             }
 
             if(body.contains("pinnedservercert")) {
